corrigido função pot.c

+++ src/functions.h
@@ -1,8 +1,8 @@
-#pragma once 
+#pragma once
 
-int add (int a, int b);
+int add(int a, int b);
 int sub(int a, int b);
 int mul(int a, int b);
 double divi(double a, double b);
 double raiz(double a);
-int pot(int a, int b);
+double pot(double a, int b);
